+++ lib/rdf/quad.ex
@@ -6,7 +6,7 @@ defmodule RDF.Quad do
   RDF values for subject, predicate, object and a graph context.
   """
 
-  alias RDF.Statement
+  alias RDF.{Statement, Term}
 
   @doc """
   Creates a `RDF.Quad` with proper RDF values.
@@ -47,5 +47,30 @@ defmodule RDF.Quad do
   """
   def new({subject, predicate, object, graph_context}),
     do: new(subject, predicate, object, graph_context)
+
+
+  @doc """
+  Returns a tuple of native Elixir values from a `RDF.Quad` of RDF terms.
+
+  Returns `nil` if one of the components of the given tuple is not convertible via `RDF.Term.value/1`.
+
+  ## Examples
+
+      iex> RDF.Quad.values {~I<http://example.com/S>, ~I<http://example.com/p>, RDF.literal(42), ~I<http://example.com/Graph>}
+      {"http://example.com/S", "http://example.com/p", 42, "http://example.com/Graph"}
+
+  """
+  def values({subject, predicate, object, graph_context}) do
+    with subject_value   when not is_nil(subject_value)       <- Term.value(subject),
+         predicate_value when not is_nil(predicate_value)     <- Term.value(predicate),
+         object_value    when not is_nil(object_value)        <- Term.value(object),
+         graph_context_value when not is_nil(graph_context_value) or is_nil(graph_context) <-
+           Term.value(graph_context)
+    do
+      {subject_value, predicate_value, object_value, graph_context_value}
+    end
+  end
+
+  def values(_), do: nil
 
 end